тип метки времени для определения актуальности данных изменен с uint64_t на uint32_t

+++ src/meters.h
@@ -54,7 +54,7 @@ typedef struct{
     meter_parameters_t parameters;
     meters_values_t values;
     int32_t isValid;
-    uint64_t timemark;
+    uint32_t timemark;
 }meter_itemInfo_t;
 
 typedef struct{
